chore: address collection reorder review feedback
